+++ CHANGELOG.md
@@ -41,6 +41,8 @@ RELEASING:
  -->
 
 ## Unreleased
+### Added
+- Processing algorithms for the Network Export endpoint ([#210](https://github.com/GIScience/orstools-qgis-plugin/issues/210))
 
 ### Changed
 - Use QgsSettings instead of config.yml file to avoid deletion of providers on update ([#108](https://github.com/GIScience/orstools-qgis-plugin/issues/108))